pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "skill-optimizer"
version = "0.3.0"
description = "Auto-track and optimize Claude skills using AI-powered conversation analysis"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Meet Patel", email = "meet84046@gmail.com"}
]
keywords = ["claude", "anthropic", "skills", "optimization", "ai"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
requires-python = ">=3.9"
dependencies = [
    "pyyaml>=6.0",
    "python-dotenv>=1.0.0",
    "google-genai>=1.0.0",
]

[project.optional-dependencies]
anthropic = [
    "anthropic>=0.50.0",
]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "anthropic>=0.50.0",
]

[project.urls]
Repository = "https://github.com/meet-rocking/skill-optimizer"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
"*" = ["*.yaml", "*.json"]
